Development ray

+++ src/main/java/org/example/interfaces/SensorModelRepositoryJpa.java
@@ -3,5 +3,5 @@ package org.example.interfaces;
 import org.example.models.SensorModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SensorModelRepositoryJpa extends JpaRepository<SensorModel, Long> {
+public interface SensorModelRepositoryJpa extends JpaRepository<SensorModel, Integer> {
 }
